FreeTrialAd refactored decorator
